+++ internal/controller/ws/client.go
@@ -1,17 +1,15 @@
-package controller
+package ws
 
 import (
 	"context"
 	"errors"
 	"log"
-	"net/http"
 	"time"
 
+	"github.com/RedHatInsights/platform-receptor-controller/internal/controller"
 	"github.com/RedHatInsights/platform-receptor-controller/internal/platform/queue"
 	"github.com/RedHatInsights/platform-receptor-controller/internal/receptor/protocol"
-	"github.com/gorilla/mux"
 	"github.com/gorilla/websocket"
-	"github.com/redhatinsights/platform-go-middlewares/identity"
 )
 
 const (
@@ -48,12 +46,12 @@ type rcClient struct {
 	socket *websocket.Conn
 
 	// send is a channel on which messages are sent.
-	send chan Work
+	send chan controller.Work
 
 	cancel context.CancelFunc
 }
 
-func (c *rcClient) SendWork(w Work) {
+func (c *rcClient) SendWork(w controller.Work) {
 	c.send <- w
 }
 
@@ -66,67 +64,6 @@ func (c *rcClient) Close() {
 	// FIXME:  Think through this a bit more.  On close, we might need to to try
 	// send a CloseMessage to the client??
 	c.cancel()
-}
-
-func performHandshake(socket *websocket.Conn) (string, error) {
-
-	socket.SetReadDeadline(time.Now().Add(pongWait))
-
-	messageType, r, err := socket.NextReader()
-	log.Println("WebSocket reader got a message...")
-	if err != nil {
-		log.Println("WebSocket reader - error: ", err)
-		return "", err
-	}
-
-	if messageType != websocket.BinaryMessage {
-		log.Printf("WebSocket reader: invalid type, expected %d, got %d", websocket.BinaryMessage, messageType)
-		return "", errors.New("websocket reader: invalid message type")
-	}
-
-	message, err := protocol.ReadMessage(r)
-	if err != nil {
-		log.Println("Websocket reader - error reading/parsing message: ", err)
-		return "", err
-	}
-	log.Println("Websocket reader message:", message)
-	log.Println("Websocket reader message type:", message.Type())
-
-	if message.Type() != protocol.HiMessageType {
-		log.Printf("WebSocket reader: invalid type, expected %d, got %d", protocol.HiMessageType, message.Type())
-		return "", errors.New("websocket reader: invalid receptor message type")
-	}
-
-	hiMessage, ok := message.(*protocol.HiMessage)
-	if ok != true {
-		log.Println("Websocket reader - error casting message to HiMessage")
-		return "", errors.New("websocket reader: invalid receptor message type")
-	}
-
-	log.Printf("Received a hi message from receptor node %s\n", hiMessage.ID)
-
-	log.Println("WebSocket writer - sending HI")
-
-	socket.SetWriteDeadline(time.Now().Add(writeWait))
-	w, err := socket.NextWriter(websocket.BinaryMessage)
-	if err != nil {
-		log.Println("WebSocket writer - error getting next writer: ", err)
-		return "", err
-	}
-
-	defer w.Close()
-
-	responseHiMessage := protocol.HiMessage{Command: "HI", ID: receptorControllerNodeId}
-
-	err = protocol.WriteMessage(w, &responseHiMessage)
-	if err != nil {
-		log.Println("WebSocket writer - error writing message: ", err)
-		return "", err
-	}
-
-	log.Println("WebSocket writer - sent HI")
-
-	return hiMessage.ID, nil
 }
 
 func (c *rcClient) read(ctx context.Context) {
@@ -251,7 +188,7 @@ func (c *rcClient) consume(ctx context.Context) {
 
 		if string(m.Key) == c.account {
 			// FIXME:
-			w := Work{}
+			w := controller.Work{}
 			c.SendWork(w)
 		} else {
 			log.Println("Kafka job reader - received message but did not send. Account number not found.")
@@ -259,74 +196,63 @@ func (c *rcClient) consume(ctx context.Context) {
 	}
 }
 
-type ReceptorController struct {
-	connectionMgr *ConnectionManager
-	router        *mux.Router
-}
+func performHandshake(socket *websocket.Conn) (string, error) {
 
-func NewReceptorController(cm *ConnectionManager, r *mux.Router) *ReceptorController {
-	return &ReceptorController{
-		connectionMgr: cm,
-		router:        r,
+	socket.SetReadDeadline(time.Now().Add(pongWait))
+
+	messageType, r, err := socket.NextReader()
+	log.Println("WebSocket reader got a message...")
+	if err != nil {
+		log.Println("WebSocket reader - error: ", err)
+		return "", err
 	}
-}
 
-const (
-	socketBufferSize  = 1024
-	messageBufferSize = 256
-)
-
-var upgrader = &websocket.Upgrader{ReadBufferSize: socketBufferSize, WriteBufferSize: socketBufferSize}
-
-func (rc *ReceptorController) Routes() {
-	rc.router.HandleFunc("/wss/receptor-controller/gateway", rc.handleWebSocket())
-	rc.router.Use(identity.EnforceIdentity)
-}
-
-func (rc *ReceptorController) handleWebSocket() http.HandlerFunc {
-
-	return func(w http.ResponseWriter, req *http.Request) {
-
-		socket, err := upgrader.Upgrade(w, req, nil)
-		if err != nil {
-			log.Println("Upgrade error:", err)
-			return
-		}
-
-		rhIdentity := identity.Get(req.Context())
-
-		log.Println("WebSocket server - got a connection, account #", rhIdentity.Identity.AccountNumber)
-		log.Println("All the headers: ", req.Header)
-
-		client := &rcClient{
-			account: rhIdentity.Identity.AccountNumber,
-			socket:  socket,
-			send:    make(chan Work, messageBufferSize),
-		}
-
-		ctx := req.Context()
-		ctx, cancel := context.WithCancel(ctx)
-		defer cancel()
-		client.cancel = cancel
-
-		socket.SetReadLimit(maxMessageSize)
-
-		defer socket.Close()
-
-		peerID, err := performHandshake(client.socket)
-		if err != nil {
-			log.Println("Error during handshake:", err)
-			return
-		}
-
-		rc.connectionMgr.Register(client.account, peerID, client)
-
-		// once this go routine exits...notify the connection manager of the clients departure
-		defer func() {
-			rc.connectionMgr.Unregister(client.account, peerID)
-			log.Println("Websocket server - account unregistered from connection manager")
-		}()
-
-		client.read(ctx)
+	if messageType != websocket.BinaryMessage {
+		log.Printf("WebSocket reader: invalid type, expected %d, got %d", websocket.BinaryMessage, messageType)
+		return "", errors.New("websocket reader: invalid message type")
 	}
+
+	message, err := protocol.ReadMessage(r)
+	if err != nil {
+		log.Println("Websocket reader - error reading/parsing message: ", err)
+		return "", err
+	}
+	log.Println("Websocket reader message:", message)
+	log.Println("Websocket reader message type:", message.Type())
+
+	if message.Type() != protocol.HiMessageType {
+		log.Printf("WebSocket reader: invalid type, expected %d, got %d", protocol.HiMessageType, message.Type())
+		return "", errors.New("websocket reader: invalid receptor message type")
+	}
+
+	hiMessage, ok := message.(*protocol.HiMessage)
+	if ok != true {
+		log.Println("Websocket reader - error casting message to HiMessage")
+		return "", errors.New("websocket reader: invalid receptor message type")
+	}
+
+	log.Printf("Received a hi message from receptor node %s\n", hiMessage.ID)
+
+	log.Println("WebSocket writer - sending HI")
+
+	socket.SetWriteDeadline(time.Now().Add(writeWait))
+	w, err := socket.NextWriter(websocket.BinaryMessage)
+	if err != nil {
+		log.Println("WebSocket writer - error getting next writer: ", err)
+		return "", err
+	}
+
+	defer w.Close()
+
+	responseHiMessage := protocol.HiMessage{Command: "HI", ID: receptorControllerNodeId}
+
+	err = protocol.WriteMessage(w, &responseHiMessage)
+	if err != nil {
+		log.Println("WebSocket writer - error writing message: ", err)
+		return "", err
+	}
+
+	log.Println("WebSocket writer - sent HI")
+
+	return hiMessage.ID, nil
 }
